Moved shared components and put router-links in nav-bar
