+++ site/ui/router/index.js
@@ -23,7 +23,7 @@ export default new Router({
     },
     {
       path: '/contact',
-      name: 'Contact Us',
+      name: 'Contact',
       component: Components.Contact,
     },
     {
